fix: Fix empty body logic and refactor
